cleanup mdns handling
