--- client/service_client.go
@@ -1,32 +1,0 @@
-package client
-
-import (
-	"fmt"
-	"strconv"
-	"time"
-)
-
-// getIPAddressPort gets the ip address and port from the given service info object.
-// It passes the caller the ip address and the port separated by ":" together in a string.
-func getIPAddressPort(svcInfo *ServiceInfo) string {
-	ipAddress, port := svcInfo.GetNetAddress()
-	ipAddrPort := ipAddress + ":" + strconv.Itoa(int(port))
-	return ipAddrPort
-}
-
-// NewClientFromService creates a new core function client from a socket with a address, which was acquired from the specified service.
-// The timeout specifies the maximal time waiting for a service to show up.
-func NewClientFromService(serviceAddr string, timeout time.Duration) (*Client, error) {
-	instance, service, err := ParseInstanceAndService(serviceAddr)
-	if err != nil {
-		return nil, err
-	}
-	svcInfo, err := NewServiceInfo(instance, service, timeout)
-	if err != nil {
-		return nil, err
-	}
-	ipAddrPort := getIPAddressPort(svcInfo)
-	fmt.Println("Try to connect with ", ipAddrPort)
-	c, err := NewClientFromSocketAddress(ipAddrPort)
-	return c, err
-}
